extract and improve changelog component
